// src/app/shared/interfaces/Vehicle.ts
export interface Vehicle {
  id: number;
  colourOfSticker: string;
  companyId: number;
  companyNumber: string;
  companyName: string;
  vehicleGSE: string;
  manufacturer: string;
  carRegistrationLicense: string;
  plateNumber: string;
  tires: string;
  type: string;
  category: string;
  barcode: string;
  lastInspectionDate: Date;
  lastInspectionResult: string;
}